+++ addons/pt/saft/scenarios_test.go
@@ -54,23 +54,50 @@ func TestInvoice(t *testing.T) {
 		assert.NoError(t, inv.Validate())
 	})
 
+	t.Run("exempt with addon added later", func(t *testing.T) {
+		// This tests covers a typical use-case whereby a document is
+		// created without addons but with the extensions to be used later.
+		inv := validInvoice()
+		inv.Addons = tax.Addons{}
+		tc := inv.Lines[0].Taxes[0]
+		tc.Key = ""
+		tc.Rate = tax.KeyExempt
+		tc.Ext = tc.Ext.Set(saft.ExtKeyExemption, "M40")
+
+		require.NoError(t, inv.Calculate())
+
+		assert.Empty(t, tc.Ext[saft.ExtKeyTaxRate].String())
+		assert.Equal(t, "exempt", tc.Key.String())
+		assert.Equal(t, "M40", tc.Ext[saft.ExtKeyExemption].String())
+
+		// Add the addon and re-calculate
+		inv.Addons = tax.WithAddons(saft.V1)
+		require.NoError(t, inv.Calculate())
+		assert.Equal(t, "ISE", tc.Ext[saft.ExtKeyTaxRate].String())
+		assert.Equal(t, "reverse-charge", tc.Key.String())
+		assert.Equal(t, "M40", tc.Ext[saft.ExtKeyExemption].String())
+	})
+
 	t.Run("exempt", func(t *testing.T) {
 		inv := validInvoice()
 		tc := inv.Lines[0].Taxes[0]
 		tc.Key = tax.KeyExempt
 		tc.Rate = ""
 		require.NoError(t, inv.Calculate())
-		assert.Equal(t, "ISE", inv.Lines[0].Taxes[0].Ext[saft.ExtKeyTaxRate].String())
-		assert.Equal(t, "M07", inv.Lines[0].Taxes[0].Ext[saft.ExtKeyExemption].String())
+		assert.Equal(t, "ISE", tc.Ext[saft.ExtKeyTaxRate].String())
+		assert.Equal(t, "M07", tc.Ext[saft.ExtKeyExemption].String())
 
-		// Allow override
-		inv.Lines[0].Taxes[0].Ext[saft.ExtKeyExemption] = "M04"
+		// Allow override as this is "exempt"
+		tc.Ext[saft.ExtKeyExemption] = "M04"
 		require.NoError(t, inv.Calculate())
-		assert.Equal(t, "M07", inv.Lines[0].Taxes[0].Ext[saft.ExtKeyExemption].String())
+		assert.Equal(t, "export", tc.Key.String())
+		assert.Equal(t, "M04", tc.Ext[saft.ExtKeyExemption].String())
 
-		// Allow override
-		inv.Lines[0].Taxes[0].Ext[saft.ExtKeyExemption] = "M01"
+		// Do not allow override from "export" back to "exempt", but
+		// force the code back to default "M05"
+		tc.Ext[saft.ExtKeyExemption] = "M01"
 		require.NoError(t, inv.Calculate())
-		assert.Equal(t, "M01", inv.Lines[0].Taxes[0].Ext[saft.ExtKeyExemption].String())
+		assert.Equal(t, "export", tc.Key.String())
+		assert.Equal(t, "M05", tc.Ext[saft.ExtKeyExemption].String())
 	})
 }